Condense export and index jobs into one asset group
